Update the organisation of test files
